Support JSON files larger than 2 GB

+++ src/main/java/io/georocket/util/JsonStreamEvent.java
@@ -25,7 +25,7 @@ public class JsonStreamEvent extends StreamEvent {
    * @param currentValue the value or field name that was parsed when the event
    * was generated (may be <code>null</code>)
    */
-  public JsonStreamEvent(int event, int pos, Object currentValue) {
+  public JsonStreamEvent(int event, long pos, Object currentValue) {
     super(pos);
     this.event = event;
     this.currentValue = currentValue;
